role hidden in token + refactor
